Rename NonexistentServerUrlDockerCredentialHelperException to UnknownServerUrlException, parent CredentialRetrieverException to other credential retrieval-related exceptions
