+++ jib-core/src/main/java/com/google/cloud/tools/jib/frontend/CredentialRetrieverFactory.java
@@ -19,12 +19,12 @@ package com.google.cloud.tools.jib.frontend;
 import com.google.cloud.tools.jib.JibLogger;
 import com.google.cloud.tools.jib.configuration.credentials.Credential;
 import com.google.cloud.tools.jib.configuration.credentials.CredentialRetriever;
-import com.google.cloud.tools.jib.configuration.credentials.CredentialRetriever.CredentialRetrievalException;
 import com.google.cloud.tools.jib.image.ImageReference;
+import com.google.cloud.tools.jib.registry.credentials.CredentialRetrievalException;
 import com.google.cloud.tools.jib.registry.credentials.DockerConfigCredentialRetriever;
 import com.google.cloud.tools.jib.registry.credentials.DockerCredentialHelperFactory;
 import com.google.cloud.tools.jib.registry.credentials.DockerCredentialHelperNotFoundException;
-import com.google.cloud.tools.jib.registry.credentials.NonexistentServerUrlDockerCredentialHelperException;
+import com.google.cloud.tools.jib.registry.credentials.UnknownServerUrlException;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.ImmutableMap;
 import java.io.IOException;
@@ -181,7 +181,7 @@ public class CredentialRetrieverFactory {
             }
           }
 
-        } catch (NonexistentServerUrlDockerCredentialHelperException | IOException ex) {
+        } catch (UnknownServerUrlException | IOException ex) {
           throw new CredentialRetrievalException(ex);
         }
       }
@@ -198,7 +198,7 @@ public class CredentialRetrieverFactory {
       try {
         return retrieveFromDockerCredentialHelper(credentialHelper, dockerCredentialHelperFactory);
 
-      } catch (NonexistentServerUrlDockerCredentialHelperException ex) {
+      } catch (UnknownServerUrlException ex) {
         logger.info(
             "No credentials for " + imageReference.getRegistry() + " in " + credentialHelper);
         return null;
@@ -229,8 +229,7 @@ public class CredentialRetrieverFactory {
 
   private Credential retrieveFromDockerCredentialHelper(
       Path credentialHelper, DockerCredentialHelperFactory dockerCredentialHelperFactory)
-      throws NonexistentServerUrlDockerCredentialHelperException,
-          DockerCredentialHelperNotFoundException, IOException {
+      throws UnknownServerUrlException, DockerCredentialHelperNotFoundException, IOException {
     Credential credentials =
         dockerCredentialHelperFactory
             .newDockerCredentialHelper(imageReference.getRegistry(), credentialHelper)
