+++ jib-core/src/test/java/com/google/cloud/tools/jib/frontend/CredentialRetrieverFactoryTest.java
@@ -23,7 +23,7 @@ import com.google.cloud.tools.jib.registry.credentials.DockerConfigCredentialRet
 import com.google.cloud.tools.jib.registry.credentials.DockerCredentialHelper;
 import com.google.cloud.tools.jib.registry.credentials.DockerCredentialHelperFactory;
 import com.google.cloud.tools.jib.registry.credentials.DockerCredentialHelperNotFoundException;
-import com.google.cloud.tools.jib.registry.credentials.NonexistentServerUrlDockerCredentialHelperException;
+import com.google.cloud.tools.jib.registry.credentials.UnknownServerUrlException;
 import java.io.IOException;
 import java.nio.file.Paths;
 import org.junit.Assert;
@@ -54,8 +54,7 @@ public class CredentialRetrieverFactoryTest {
 
   @Before
   public void setUp()
-      throws NonexistentServerUrlDockerCredentialHelperException,
-          DockerCredentialHelperNotFoundException, IOException {
+      throws UnknownServerUrlException, DockerCredentialHelperNotFoundException, IOException {
     Mockito.when(mockDockerCredentialHelper.retrieve()).thenReturn(FAKE_CREDENTIALS);
     Mockito.when(mockNonexistentDockerCredentialHelper.retrieve())
         .thenThrow(mockDockerCredentialHelperNotFoundException);
