+++ jib-core/src/test/java/com/google/cloud/tools/jib/registry/credentials/DockerConfigCredentialRetrieverTest.java
@@ -43,8 +43,8 @@ public class DockerConfigCredentialRetrieverTest {
 
   @Before
   public void setUp()
-      throws URISyntaxException, NonexistentServerUrlDockerCredentialHelperException,
-          DockerCredentialHelperNotFoundException, IOException {
+      throws URISyntaxException, UnknownServerUrlException, DockerCredentialHelperNotFoundException,
+          IOException {
     dockerConfigFile = Paths.get(Resources.getResource("json/dockerconfig.json").toURI());
 
     Mockito.when(mockDockerCredentialHelper.retrieve()).thenReturn(FAKE_CREDENTIAL);
